ECNG-1552: fix

+++ EasyCard/Data/Transactions.Database/dbo/Tables/MasavFileRow.sql
@@ -3,19 +3,21 @@
     [MasavFileID]          BIGINT           NULL,
     [PaymentTransactionID] UNIQUEIDENTIFIER NULL,
     [ConsumerID]           UNIQUEIDENTIFIER NULL,
-    [Bankcode]             INT              NULL,
-    [BranchNumber]         INT              NULL,
-    [AccountNumber]        INT              NULL,
-    [NationalID]           INT              NULL,
+    [Bankcode]             BIGINT           NULL,
+    [BranchNumber]         BIGINT           NULL,
+    [AccountNumber]        BIGINT           NULL,
+    [NationalID]           BIGINT           NULL,
     [Amount]               DECIMAL (19, 4)  NULL,
     [IsPayed]              BIT              NULL,
     [SmsSent]              BIT              NOT NULL,
     [SmsSentDate]          DATETIME2 (7)    NULL,
     [ConsumerName]         NVARCHAR (50)    NULL,
-    [InstituteNumber]      INT              NULL,
+    [InstituteNumber]      BIGINT           NULL,
     CONSTRAINT [PK_MasavFileRow] PRIMARY KEY CLUSTERED ([MasavFileRowID] ASC),
     CONSTRAINT [FK_MasavFileRow_MasavFile_MasavFileID] FOREIGN KEY ([MasavFileID]) REFERENCES [dbo].[MasavFile] ([MasavFileID])
 );
+
+
